backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cleo==2.1.0
constantly==23.10.4
crashtest==0.4.1
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.0
django-filter==23.3
django-formtools==2.4.1
django-oauth-toolkit==2.3.0
django-otp==1.3.0
django-phonenumber-field==7.2.0
django-two-factor-auth==1.15.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dulwich==0.21.6
fastjsonschema==2.19.0
filelock==3.13.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
jwcrypto==1.5.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.5.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
SecretStorage==3.3.3
service-identity==23.1.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
tqdm==4.66.1
trove-classifiers==2023.11.22
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.7
wheel==0.41.3
wrapt==1.16.0
zope.interface==6.1
